store/tikv: move client*.go into single package tikv/client
