+++ store/tikv/client/mock_tikv_service_test.go
@@ -1,4 +1,4 @@
-package tikv
+package client
 
 import (
 	"context"
